requirements.txt
numpy
pandas
pillow
streamlit
tensorflow==2.18.0